# specs/001-ai/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: 智能標書產生器系統 API
  description: 提供公司資料管理、範本管理、AI內容生成和標書編輯匯出功能的RESTful API
  version: 1.0.0
  contact:
    name: 標書產生器團隊
    email: support@proposal-generator.com
servers:
  - url: http://localhost:3000/api/v1
    description: 開發環境
  - url: https://api.proposal-generator.com/v1
    description: 生產環境
paths:
...
  /templates:
    get:
      tags: [Templates]
      summary: 取得標書範本列表
      security:
        - bearerAuth: []
      parameters:
        - name: template_type
          in: query
          schema:
            $ref: '#/components/schemas/TemplateType'
      responses:
        '200':
          description: 成功取得範本列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProposalTemplate'

    post:
      tags: [Templates]
      summary: 建立自訂範本
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
      responses:
        '201':
          description: 成功建立範本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalTemplate'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProposalTemplate:
      type: object
      properties:
        id:
          type: string
        template_name:
          type: string
        template_type:
          $ref: '#/components/schemas/TemplateType'
        description:
          type: string
        is_system_template:
          type: boolean
        created_at:
          type: string
          format: date-time
    TemplateType:
      type: string
      enum: [government_grant, enterprise_bid, award_application, custom]
...
    TemplateCreate:
      type: object
      required: [template_name, template_type]
      properties:
        template_name:
          type: string
          minLength: 2
          maxLength: 255
        template_type:
          $ref: '#/components/schemas/TemplateType'
        description:
          type: string
...